% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tcga_data.r
\name{fetch_tcga_data}
\alias{fetch_tcga_data}
\title{This function is used to fetch the TCGA data using TCGABiolink package}
\usage{
fetch_tcga_data(
  tcga_type,
  download_location,
  data_type = "Gene Expression Quantification",
  workflow_type = "HTSeq - FPKM-UQ",
  sample_type = "Primary Tumor"
)
}
\arguments{
\item{tcga_type}{TCGA cancer type to download}

\item{download_location}{location of download file}

\item{data_type}{default "Gene Expression Quantification"}

\item{workflow_type}{"HTSeq - FPKM-UQ"}

\item{sample_type}{"Primary Tumor"}
}
\value{
GDCprepare objects that can be further used for get exp matrix
}
\description{
This function is used to fetch the TCGA data using TCGABiolink package
}
