% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.r
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{title of the graph}
\usage{
grid_arrange_shared_legend(
  ...,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right"),
  title = NULL
)
}
\description{
This function is used to merge multiple ggplot2 object together and share same legend.
}
