% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.r
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{This function is used to merge multiple ggplot2 object together and share same legend.}
\usage{
grid_arrange_shared_legend(
  ...,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right"),
  title = NULL
)
}
\arguments{
\item{...}{ggplot2 objects}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{position}{"bottom" or "right", location to place the legend}

\item{title}{title of the graph}
}
\description{
This function is used to merge multiple ggplot2 object together and share same legend.
}
